requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Jinja2==3.1.3
gunicorn==21.2.0
python-dotenv==1.0.1